Move pandas scan to separate files
